Simplify abs2rel.t.
Factor out all the things that were common for the different test
cases (which was everything except the filename).  And use made-up
names that indicate what's being tested rather than special system
filenames commonly found on Unix systems.  Otherwise, if, for
example, a test for "'init.d' is a directory" fails, you'd be
wondering whether something is wrong with your system rather than
looking for problems in File::Spec, which is what we're testing.